Squashed commit of the following:
commit da0931c7339f0ac9bed37f17059c0c4f2a05aa74
Author: Xavian Escamilla <xzillagaming@gmail.com>
Date:   Mon Mar 18 08:45:31 2024 -0500

    Pushing of Diesel and Interaction Logic

    Previous Push did not contain the Diesel and Interaction scripts

commit b8b189dcc677d2675673abe3b0dc8bf5f7727fc6
Author: Xavian Escamilla <xzillagaming@gmail.com>
Date:   Mon Mar 18 08:42:42 2024 -0500

    Camera Collider, Player Gravity, & Interactable Added

    Camera now collides and slides along objects on the collision layer "Default". Player gravity has been added and Revised Player Controller script has been renamed to Player Controller and attached to player character. Basic interactable system has been added with a diesel refuel script using the interactable script.

commit ea43c4b2d029865c8a42ac40ad733d188b97aa81
Author: Xavian Escamilla <xzillagaming@gmail.com>
Date:   Wed Mar 13 12:01:46 2024 -0500

    Implented Basic Equipables and Player Movement

    System added for player to be able to have only one tool drain diesel at a time. Player Movement Revision is roughly 80% complete.

commit ddbd1205c9fd1c8a899c824acfe6bb4815facf01
Author: Xavian Escamilla <xzillagaming@gmail.com>
Date:   Wed Mar 13 09:33:30 2024 -0500

    Added Init Revised Player Movement

    Added in a Revised player movement system using the character controller instead of a Rigidbody. Cleanup of code and components to follow.

+++ Assets/Assets/Scripts/Enemy/GenericEnemy.cs
@@ -1,14 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 
 public class GenericEnemy : MonoBehaviour
 {
 
-    [SerializeField] private int health;
+    [SerializeField] private float health;
 
-    public void TakeMyDamage(int damage_taken)
+    public void TakeMyDamage(float damage_taken)
     {
         health -= damage_taken;
         Debug.Log("In Generic Enemy Script of Object" + this.gameObject.name + " " + health);
